requirements.txt
streamlit==1.37.0
langchain==0.2.11
python-dotenv==1.0.1
pillow==10.3.0
pandas==2.2.2
requests==2.32.3
google-generativeai==0.5.2


